+++ packages/tailwindcss/src/candidate.test.ts
@@ -2,7 +2,7 @@ import { expect, it } from 'vitest'
 import { buildDesignSystem } from './design-system'
 import { Theme } from './theme'
 import { Utilities } from './utilities'
-import { Variants } from './variants'
+import { Compounds, Variants } from './variants'
 
 function run(
   candidate: string,
@@ -109,7 +109,6 @@ it('should parse a simple utility with a variant', () => {
         "root": "flex",
         "variants": [
           {
-            "compounds": true,
             "kind": "static",
             "root": "hover",
           },
@@ -137,12 +136,10 @@ it('should parse a simple utility with stacked variants', () => {
         "root": "flex",
         "variants": [
           {
-            "compounds": true,
             "kind": "static",
             "root": "hover",
           },
           {
-            "compounds": true,
             "kind": "static",
             "root": "focus",
           },
@@ -166,7 +163,6 @@ it('should parse a simple utility with an arbitrary variant', () => {
         "root": "flex",
         "variants": [
           {
-            "compounds": true,
             "kind": "arbitrary",
             "relative": false,
             "selector": "& p",
@@ -194,7 +190,6 @@ it('should parse a simple utility with a parameterized variant', () => {
         "root": "flex",
         "variants": [
           {
-            "compounds": true,
             "kind": "functional",
             "modifier": null,
             "root": "data",
@@ -214,7 +209,7 @@ it('should parse compound variants with an arbitrary value as an arbitrary varia
   utilities.static('flex', () => [])
 
   let variants = new Variants()
-  variants.compound('group', () => {})
+  variants.compoundWith('group', Compounds.StyleRules, () => {})
 
   expect(run('group-[&_p]/parent-name:flex', { utilities, variants })).toMatchInlineSnapshot(`
     [
@@ -226,7 +221,6 @@ it('should parse compound variants with an arbitrary value as an arbitrary varia
         "root": "flex",
         "variants": [
           {
-            "compounds": true,
             "kind": "compound",
             "modifier": {
               "kind": "named",
@@ -234,7 +228,6 @@ it('should parse compound variants with an arbitrary value as an arbitrary varia
             },
             "root": "group",
             "variant": {
-              "compounds": true,
               "kind": "arbitrary",
               "relative": false,
               "selector": "& p",
@@ -251,7 +244,7 @@ it('should parse a simple utility with a parameterized variant and a modifier', 
   utilities.static('flex', () => [])
 
   let variants = new Variants()
-  variants.compound('group', () => {})
+  variants.compoundWith('group', Compounds.StyleRules, () => {})
   variants.functional('aria', () => {})
 
   expect(run('group-aria-[disabled]/parent-name:flex', { utilities, variants }))
@@ -265,7 +258,6 @@ it('should parse a simple utility with a parameterized variant and a modifier', 
           "root": "flex",
           "variants": [
             {
-              "compounds": true,
               "kind": "compound",
               "modifier": {
                 "kind": "named",
@@ -273,7 +265,6 @@ it('should parse a simple utility with a parameterized variant and a modifier', 
               },
               "root": "group",
               "variant": {
-                "compounds": true,
                 "kind": "functional",
                 "modifier": null,
                 "root": "aria",
@@ -295,7 +286,7 @@ it('should parse compound group with itself group-group-*', () => {
 
   let variants = new Variants()
   variants.static('hover', () => {})
-  variants.compound('group', () => {})
+  variants.compoundWith('group', Compounds.StyleRules, () => {})
 
   expect(run('group-group-group-hover/parent-name:flex', { utilities, variants }))
     .toMatchInlineSnapshot(`
@@ -308,7 +299,6 @@ it('should parse compound group with itself group-group-*', () => {
           "root": "flex",
           "variants": [
             {
-              "compounds": true,
               "kind": "compound",
               "modifier": {
                 "kind": "named",
@@ -316,17 +306,14 @@ it('should parse compound group with itself group-group-*', () => {
               },
               "root": "group",
               "variant": {
-                "compounds": true,
                 "kind": "compound",
                 "modifier": null,
                 "root": "group",
                 "variant": {
-                  "compounds": true,
                   "kind": "compound",
                   "modifier": null,
                   "root": "group",
                   "variant": {
-                    "compounds": true,
                     "kind": "static",
                     "root": "hover",
                   },
@@ -353,7 +340,6 @@ it('should parse a simple utility with an arbitrary media variant', () => {
         "root": "flex",
         "variants": [
           {
-            "compounds": true,
             "kind": "arbitrary",
             "relative": false,
             "selector": "@media(width>=123px)",
@@ -478,7 +464,6 @@ it('should parse a utility with a modifier and a variant', () => {
         },
         "variants": [
           {
-            "compounds": true,
             "kind": "static",
             "root": "hover",
           },
@@ -895,7 +880,6 @@ it('should parse a static variant starting with @', () => {
         "root": "flex",
         "variants": [
           {
-            "compounds": true,
             "kind": "static",
             "root": "@lg",
           },
@@ -922,7 +906,6 @@ it('should parse a functional variant with a modifier', () => {
         "root": "flex",
         "variants": [
           {
-            "compounds": true,
             "kind": "functional",
             "modifier": {
               "kind": "named",
@@ -957,7 +940,6 @@ it('should parse a functional variant starting with @', () => {
         "root": "flex",
         "variants": [
           {
-            "compounds": true,
             "kind": "functional",
             "modifier": null,
             "root": "@",
@@ -989,7 +971,6 @@ it('should parse a functional variant starting with @ and a modifier', () => {
         "root": "flex",
         "variants": [
           {
-            "compounds": true,
             "kind": "functional",
             "modifier": {
               "kind": "named",
@@ -1204,7 +1185,6 @@ it('should parse arbitrary properties with a variant', () => {
         "value": "red",
         "variants": [
           {
-            "compounds": true,
             "kind": "static",
             "root": "hover",
           },
@@ -1230,12 +1210,10 @@ it('should parse arbitrary properties with stacked variants', () => {
         "value": "red",
         "variants": [
           {
-            "compounds": true,
             "kind": "static",
             "root": "hover",
           },
           {
-            "compounds": true,
             "kind": "static",
             "root": "focus",
           },
@@ -1257,13 +1235,11 @@ it('should parse arbitrary properties that are important and using stacked arbit
         "value": "red",
         "variants": [
           {
-            "compounds": true,
             "kind": "arbitrary",
             "relative": false,
             "selector": "& p",
           },
           {
-            "compounds": true,
             "kind": "arbitrary",
             "relative": false,
             "selector": "@media(width>=123px)",
@@ -1279,7 +1255,7 @@ it('should not parse compound group with a non-compoundable variant', () => {
   utilities.static('flex', () => [])
 
   let variants = new Variants()
-  variants.compound('group', () => {})
+  variants.compoundWith('group', Compounds.StyleRules, () => {})
 
   expect(run('group-*:flex', { utilities, variants })).toMatchInlineSnapshot(`[]`)
 })
@@ -1301,7 +1277,6 @@ it('should parse a variant containing an arbitrary string with unbalanced parens
         "root": "flex",
         "variants": [
           {
-            "compounds": true,
             "kind": "functional",
             "modifier": null,
             "root": "string",
@@ -1349,7 +1324,6 @@ it('should parse candidates with a prefix', () => {
         "root": "flex",
         "variants": [
           {
-            "compounds": true,
             "kind": "static",
             "root": "hover",
           },